Checkboxes, and a bug discovered.
Most widgets are implemented with the assumption that a call to `paint` is always preceded by a call to `layout`. However, a **speculative** call to layout may happen in between. With the default `LayoutCache` behavior being to invalidate the cached value and **not** replacing it during speculative layouts, there was an issue with a field of StyledBox which was cleared during layout, and accessed afterwards in paint, resulting in a panic.

Explaining the precise sequence of calls to layout and paint is becoming tricky, but understanding it is necessary avoiding crashes like this. Furthermore, nothing of this is enforced by the compiler. So a redesign may be necessary here.