Fix tracking of hovered & hot widgets.
Previously, hovered & hot widget tracking was done in WidgetPod, and widgets with their own identity not wrapped in a WidgetPod would not be added to the hover set.

Now hover tracking is done in `core::do_event`. In order for that to work correctly, widgets that perform a hit-test should set `EventCtx::hit_test_pass` to false when the hit-test fails.

Also:
* pointer events now have an optional target, so that PointerOver/Out/Enter/Exit are only delivered to the intended widgets
* drawable widget: fixed size widget that calls a paint callback
* checkbox: draw check mark with skia & proper dark theme
* debug: show IDs of widgets (TODO: polish)

TODO:
- find a simple format for parameterized vector graphics: would be used to draw simple UI elements that adapt to the theme environment, like the check mark in a checkbox.
There's a surprising lack of good, widely-used solutions for that. SVG is too complicated to parse and doesn't expose parameters.