Data-driven tables
Experiment with tables that generate their rows & cells via traits instead of letting the application explicitly create row and column objects.

The main trait is `table::Collection` which defines a tree data model on a type. The associated type `Item` defines the row type. Each column has an associated `Fn(&Item) -> Arc<WidgetPod>` that produces the widget to put in the column , given a reference to a row.

Also:
* renamed `cache::enter` to `cache::enter_index`, `cache::enter` now takes an arbitrary hashable value.

* added `WidgetExt::arc_dyn_pod` as a convenience when rustc doesn't want to coerce `Arc<WidgetPod<impl Widget>>` to `Arc<WidgetPod<dyn Widget>>`